backendpicpay/requirements.txt
Django
djangorestframework
python-dotenv
django-cors-headers